ASL_CNN_Project/requirements.txt
opencv-python
mediapipe
numpy
tensorflow==2.12.0
Pillow
pyttsx3